Refactor InMemoryDynamoDBCompositePrimaryKeyTable into several extensions.
