+++ Sources/DynamoDBTables/InMemoryDynamoDBCompositePrimaryKeyTable+transform.swift
@@ -2,16 +2,6 @@
 //
 // This source file is part of the DynamoDBTables open source project
 //
-// This file is forked from
-// https://github.com/amzn/smoke-dynamodb/tree/main/Sources/SmokeDynamoDB/InMemoryDynamoDBCompositePrimaryKeyTableStore.swift
-// Copyright 2018-2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// Licensed under Apache License v2.0
-//
-// Changes specified by
-// https://github.com/swift-server-community/dynamo-db-tables/compare/9ab0e7a..main
-// Copyright (c) 2024 the DynamoDBTables authors
-// Licensed under Apache License v2.0
-//
 // See LICENSE.txt for license information
 // See CONTRIBUTORS.txt for the list of DynamoDBTables authors
 //
@@ -20,7 +10,7 @@
 //===----------------------------------------------------------------------===//
 
 //
-//  InMemoryDynamoDBCompositePrimaryKeyTableStore+support.swift
+//  InMemoryDynamoDBCompositePrimaryKeyTable+transform.swift
 //  DynamoDBTables
 //
 
@@ -467,5 +457,13 @@ extension InMemoryDynamoDBCompositePrimaryKeyTable {
         }
 
         return nil
+    }
+
+    func convertToQueryableType<ReturnedType: PolymorphicOperationReturnType>(input: InMemoryDatabaseItem) throws -> ReturnedType {
+        let attributeValue = DynamoDBClientTypes.AttributeValue.m(input.item)
+
+        let decodedItem: ReturnTypeDecodable<ReturnedType> = try DynamoDBDecoder().decode(attributeValue)
+
+        return decodedItem.decodedValue
     }
 }
