+++ Sources/DynamoDBTables/InMemoryDynamoDBCompositePrimaryKeyTableStore.swift
@@ -40,54 +40,4 @@ actor InMemoryDynamoDBCompositePrimaryKeyTableStore {
     func execute<Result>(_ operation: (inout StoreType) throws -> Result) rethrows -> Result {
         try operation(&self.store)
     }
-    /*
-     nonisolated func validateEntry(entry: WriteEntry<some Any, some Any, some Any>) throws {
-         let entryString = "\(entry)"
-         if entryString.count > AWSDynamoDBLimits.maxStatementLength {
-             throw DynamoDBTableError.statementLengthExceeded(
-                 reason: "failed to satisfy constraint: Member must have length less than or equal to "
-                     + "\(AWSDynamoDBLimits.maxStatementLength). Actual length \(entryString.count)")
-         }
-     }
-
-     func polymorphicBulkWrite(
-         _ entries: [some PolymorphicWriteEntry], constraints: [some PolymorphicTransactionConstraintEntry],
-         isTransaction: Bool) throws
-     {
-         let entryCount = entries.count + constraints.count
-         let context = StandardPolymorphicWriteEntryContext<InMemoryPolymorphicWriteEntryTransform,
-             InMemoryPolymorphicTransactionConstraintTransform>(table: self)
-
-         if isTransaction, entryCount > AWSDynamoDBLimits.maximumUpdatesPerTransactionStatement {
-             throw DynamoDBTableError.transactionSizeExceeded(attemptedSize: entryCount,
-                                                              maximumSize: AWSDynamoDBLimits.maximumUpdatesPerTransactionStatement)
-         }
-
-         let store = self.store
-
-         if let error = self.handleConstraints(constraints: constraints, isTransaction: isTransaction, context: context) {
-             throw error
-         }
-
-         if let error = self.handlePolymorphicEntries(entries: entries, isTransaction: isTransaction, context: context) {
-             if isTransaction {
-                 // restore the state prior to the transaction
-                 self.store = store
-             }
-
-             throw error
-         }
-     }
-
-     func deleteItems(forKeys keys: [CompositePrimaryKey<some Any>]) throws {
-         try keys.forEach { key in
-             try self.deleteItem(forKey: key)
-         }
-     }
-
-     func deleteItems<ItemType: DatabaseItem>(existingItems: [ItemType]) throws {
-         try existingItems.forEach { (existingItem: ItemType) in
-             try self.deleteItem(existingItem: existingItem)
-         }
-     }*/
 }
